+++ src/features/asset/CreateAssetModal.tsx
@@ -1,12 +1,6 @@
 import { useModal } from '@/features/hooks/useModal'
-import { Button, Container } from '@chakra-ui/react'
-import { useFormState } from 'react-dom'
-
-type FormDataType = {
-  id: FormDataEntryValue | null
-  name: FormDataEntryValue | null
-  type: FormDataEntryValue | null
-}
+import { Button } from '@chakra-ui/react'
+import { CreateAssetForm } from '@/features/asset'
 
 /**
  * 資産の登録モーダル
@@ -15,42 +9,13 @@ type FormDataType = {
 export const CreateAssetModal = () => {
   const { onOpen, ModalContainer } = useModal()
 
-  const initialState = {
-    id: '',
-    name: '',
-    type: '',
-  }
-
-  const action = (state: FormDataType, formData: FormData) => {
-    const newState = {
-      id: formData.get('id'),
-      name: formData.get('name'),
-      type: formData.get('type'),
-    }
-    return newState
-  }
-  const [, formAction] = useFormState(action, initialState)
-
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault()
-    const formData = new FormData(e.currentTarget)
-    formAction(formData)
-  }
-
   return (
     <>
       <Button colorScheme="teal" onClick={onOpen}>
         資産の登録
       </Button>
       <ModalContainer title="資産の登録">
-        <form onSubmit={handleSubmit}>
-          <Container display="flex" flexDirection="column" w={'full'} gap={4}>
-            <input type="text" name="id" placeholder="資産ID" />
-            <input type="text" name="name" placeholder="資産名" />
-            <input type="text" name="type" placeholder="資産タイプ" />
-            <Button type="submit">Submit</Button>
-          </Container>
-        </form>
+        <CreateAssetForm />
       </ModalContainer>
     </>
   )
