+++ src/features/asset/index.ts
@@ -1,4 +1,5 @@
 export { AssetListContainer } from '@/features/asset/AssetListContainer'
 export { CreateAssetModal } from '@/features/asset/CreateAssetModal'
+export { CreateAssetForm } from '@/features/asset/CreateAssetForm'
 export { AssetCardList } from '@/features/asset/AssetCardList'
 export { AssetCard } from '@/features/asset/AssetCard'
